core: Move dce to rewriter
